Extract locationiq service
